[canaria] AppComponent で管理していた auth ロジックを HomeComponent に委譲
