+++ projects/canaria/src/app/app.component.html
@@ -1,13 +1,5 @@
 <fl-header [appTitle]="appTitle"></fl-header>
 
 <main class="main">
-  <div *ngIf="user$ | async as user; else showLogin">
-    <h1>Hello {{ user.displayName }}!</h1>
-    <button (click)="logout()">Logout</button>
-  </div>
-  <ng-template #showLogin>
-    <p>Please login.</p>
-    <button (click)="login()">Login with Google</button>
-  </ng-template>
   <router-outlet></router-outlet>
 </main>
