+++ projects/canaria/src/app/pages/home/home.component.html
@@ -1,1 +1,8 @@
-<p>home works!</p>
+<div *ngIf="user$ | async as user; else showLogin">
+  <h1>Hello {{ user.displayName }}!</h1>
+  <button (click)="logout()">Logout</button>
+</div>
+<ng-template #showLogin>
+  <p>Please login.</p>
+  <button (click)="login()">Login with Google</button>
+</ng-template>
