+++ projects/canaria/src/app/pages/home/home.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { Authenticator } from 'utilities';
 
 @Component({
   selector: 'app-home',
@@ -6,7 +7,16 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./home.component.scss'],
 })
 export class HomeComponent implements OnInit {
-  constructor() {}
+  constructor(private authenticator: Authenticator) {}
+  readonly appTitle = 'Canaria';
+  user$ = this.authenticator.loggedInUser$;
+
+  login() {
+    this.authenticator.login();
+  }
+  logout() {
+    this.authenticator.logout();
+  }
 
   ngOnInit() {}
 }
